backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv
groq
httpx
python-whois
dnspython
gunicorn
